+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # dartlang plugin changelog
 
+## unreleased
+- Fixed an issue with auto-detecting the Flutter and Dart SDKs
+
 ## 0.6.3
 - Run Flutter apps with `flutter run` now instead of `flutter start`
 - Fixed an exception from the Find References view